refacto: misc and clean (#146)
* refactor: tidy, move and rename allinone

* refactor: rename and move oneline

* refactor: delete css and replace single error with toast

* chore: tidy linter pt 1

* chore: tidy linter pt 2

- fix some stuff from preview fetchapi refactor

* chore: tidy up linter (cheats)

* ci: fix production build

---------
