+++ src/AppRoutes.tsx
@@ -3,7 +3,7 @@ import { Routes, Route, Navigate } from "react-router-dom";
 import { RootLayout } from "./components/Layout/RootLayout";
 import { WalletFlow } from "./containers/WalletFlow";
 import { RequireUnlockedWallet } from "./containers/RequireUnlockedWallet";
-import { OneLiner } from "./OneLiner";
+import { MessengerContainer } from "./containers/MessengerContainer";
 import { SettingsPage } from "./SettingsPage";
 import type { NetworkType } from "./types/all";
 import type { Step } from "./containers/WalletFlow";
@@ -75,9 +75,9 @@ export const AppRoutes: React.FC<AppRoutesProps> = ({
           ))}
         </Route>
 
-        {/* Main Oneliner once you are unlocked */}
+        {/* Main Messenging container once you are unlocked */}
         <Route element={<RequireUnlockedWallet />}>
-          <Route path=":walletId" element={<OneLiner />} />
+          <Route path=":walletId" element={<MessengerContainer />} />
         </Route>
         <Route path="settings-network" element={<SettingsPage />} />
       </Route>
