--- src/components/FetchApiMessages.css
@@ -1,12 +1,0 @@
-.api-error-tooltip {
-  position: absolute;
-  bottom: -30px;
-  left: 0;
-  background-color: #ef4444;
-  color: white;
-  padding: 5px 10px;
-  border-radius: 4px;
-  font-size: 12px;
-  white-space: nowrap;
-  z-index: 10;
-}
