+++ src/utils/debug-commands.ts
@@ -35,4 +35,5 @@ export function inspectConversations() {
 }
 
 // Add to window for console access
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 (window as any).inspectConversations = inspectConversations;
